Optimization attempt
An attempt to reduce draw calls by combining tiny meshes to one, bigger mesh.
This attempt partially worked for the floor.
(Note that MeshCombiner works on it's children and is obsolete already, MeshCluster is the version I used in the end.)
Also, floors and walls are now spawned at the end. This fixes the fact that walls get spawned and then removed as the route is determined, preventing an unnecessary Instantiate and Destroy